+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Fixed:
 * Image size can now be changed in the Gutenberg image block for fully offloaded images
+* Full size images not replaced in the gallery block on expand 
 
 = 1.8.0 - 16.02.2024 =